requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

torch
torchvision
timm
pillow
numpy

prometheus-client==0.20.0
loguru==0.7.2